test(gateway): isolate hosted Runs and use signed approval scope
Move the coherent TestHostedRoomRuns class into its focused owner without duplicate collection. Replace only the exact-approval fixture auth bypass with a real invitation grant, the shared Runs scope codec and a durable reservation. Preserve all eight original assertions and the other eleven method bodies. No runtime/auth/fence behavior is changed by this test-only commit.
